+++ include/ftxui/component/component_options.hpp
@@ -72,6 +72,7 @@ struct AnimatedColorsOption {
 /// @brief Option for the MenuEntry component.
 /// @ingroup component
 struct MenuEntryOption {
+  ConstStringRef label = "MenuEntry";
   std::function<Element(const EntryState& state)> transform;
   AnimatedColorsOption animated_colors;
 };
@@ -86,9 +87,12 @@ struct MenuOption {
   static MenuOption VerticalAnimated();
   static MenuOption Toggle();
 
+  ConstStringListRef entries;  ///> The list of entries.
+  Ref<int> selected = 0;       ///> The index of the selected entry.
+
   // Style:
   UnderlineOption underline;
-  MenuEntryOption entries;
+  MenuEntryOption entries_option;
   Direction direction = Direction::Down;
   std::function<Element()> elements_prefix;
   std::function<Element()> elements_infix;
@@ -115,6 +119,9 @@ struct ButtonOption {
                                Color background_active,
                                Color foreground_active);
 
+  ConstStringRef label = "Button";
+  std::function<void()> on_click = [] {};
+
   // Style:
   std::function<Element(const EntryState&)> transform;
   AnimatedColorsOption animated_colors;
@@ -125,6 +132,10 @@ struct ButtonOption {
 struct CheckboxOption {
   // Standard constructors:
   static CheckboxOption Simple();
+
+  ConstStringRef label = "Checkbox";
+
+  Ref<bool> checked = false;
 
   // Style:
   std::function<Element(const EntryState&)> transform;
@@ -152,8 +163,9 @@ struct InputOption {
   static InputOption Default();
   /// @brief A white on black style with high margins:
   static InputOption Spacious();
-  /// @brief A style with a border:
-  static InputOption Arthur();
+
+  /// The content of the input.
+  StringRef content = "";
 
   /// The content of the input when it's empty.
   StringRef placeholder = "";
@@ -177,6 +189,10 @@ struct InputOption {
 struct RadioboxOption {
   // Standard constructors:
   static RadioboxOption Simple();
+
+  // Content:
+  ConstStringListRef entries;
+  Ref<int> selected = 0;
 
   // Style:
   std::function<Element(const EntryState&)> transform;
@@ -208,6 +224,39 @@ struct SliderOption {
   Direction direction = Direction::Right;
   Color color_active = Color::White;
   Color color_inactive = Color::GrayDark;
+};
+
+// Parameter pack used by `WindowOptions::render`.
+struct WindowRenderState {
+  Element inner;             /// < The element wrapped inside this window.
+  const std::string& title;  /// < The title of the window.
+  bool active = false;       /// < Whether the window is the active one.
+  bool drag = false;         /// < Whether the window is being dragged.
+  bool resize = false;       /// < Whether the window is being resized.
+  bool hover_left = false;   /// < Whether the resizeable left side is hovered.
+  bool hover_right = false;  /// < Whether the resizeable right side is hovered.
+  bool hover_top = false;    /// < Whether the resizeable top side is hovered.
+  bool hover_down = false;   /// < Whether the resizeable down side is hovered.
+};
+
+// @brief Option for the `Window` component.
+// @ingroup component
+struct WindowOptions {
+  Component inner;            /// < The component wrapped by this window.
+  ConstStringRef title = "";  /// < The title displayed by this window.
+
+  Ref<int> left = 0;     /// < The left side position of the window.
+  Ref<int> top = 0;      /// < The top side position of the window.
+  Ref<int> width = 20;   /// < The width of the window.
+  Ref<int> height = 10;  /// < The height of the window.
+
+  Ref<bool> resize_left = true;   /// < Can the left side be resized?
+  Ref<bool> resize_right = true;  /// < Can the right side be resized?
+  Ref<bool> resize_top = true;    /// < Can the top side be resized?
+  Ref<bool> resize_down = true;   /// < Can the down side be resized?
+
+  /// An optional function to customize how the window looks like:
+  std::function<Element(const WindowRenderState&)> render;
 };
 
 }  // namespace ftxui